chore: seperate client component from rsc with loading spinner
