+++ components/common/LoadingSpinner.tsx
@@ -1,0 +1,5 @@
+import { AiOutlineLoading } from 'react-icons/ai'
+
+export default function LoadingSpinner() {
+  return <AiOutlineLoading className="animate-spin" size={84} />
+}
